feat: implemented the worker listener and removed the Session::worker_listener method
