+++ orchestrator/src/sessions/mod.rs
@@ -1,9 +1,7 @@
-mod event;
 mod session;
 mod trained_model;
-mod training_strategies;
+mod worker_listener;
 
-pub use event::TrainingEvent;
 pub use session::{CancelHandle, Session, StopReason};
 pub use trained_model::TrainedModel;
-pub use training_strategies::{TrainingStrategy, WorkerRequest};
+pub use worker_listener::{TrainingEvent, WorkerListener, WorkerRequest};
