--- orchestrator/src/sessions/training_strategies/mod.rs
@@ -1,5 +1,0 @@
-mod all_reduce;
-mod parameter_server;
-mod strategy;
-
-pub use strategy::{TrainingStrategy, WorkerRequest};
